# PERF: Dataframe reductions with EA dtypes